Fix #60
The simple fix for getting out of sync is to only display document
highlights if the cursor is found in any of the ranges. I avoided using
this because the spec doesn't say the cursor must be in one of the
ranges. However, from my personal experience it seems that this is the
case, if I turn out to be wrong, I'll revert this which will result in
things getting out of sync when doing shit like macros.