Add a default link for the hlgroups
This will prevent `hi clear` from fully destroying the hlgroups, it will
however change the hlgroup. In reality, colorschemes should be loaded
before vim-illuminate (and plugins in general).

This change is mainly to make it obvious that the root of a filed issue
is the plugin manager (often packer.nvim), is incorrectly loading
initialization code.